+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - introduced a new command `bit untag` for reverting un-exported tags.
 - init local scope inside .git
 - support vue files
+- [#736](https://github.com/teambit/bit/issues/736) - .gitignore is blocking everything
 
 ## [0.12.5] - 2018-02-06